+++ include/votca/xtp/polarsite.h
@@ -40,8 +40,8 @@ class PolarSite : public StaticSite {
   void WriteData(StaticSite::data& d) const = delete;
   void ReadData(StaticSite::data& d) = delete;
 
-  PolarSite(int id, std::string element, Eigen::Vector3d pos);
-  PolarSite(int id, std::string element)
+  PolarSite(long id, std::string element, Eigen::Vector3d pos);
+  PolarSite(long id, std::string element)
       : PolarSite(id, element, Eigen::Vector3d::Zero()){};
 
   ~PolarSite() override = default;
@@ -88,7 +88,7 @@ class PolarSite : public StaticSite {
   }
 
   struct data {
-    int id;
+    long id;
     char* element;
     double posX;
     double posY;
