+++ include/votca/xtp/segmentmapper.h
@@ -47,18 +47,18 @@ class SegmentMapper {
   using mapAtom = typename std::iterator_traits<
       typename AtomContainer::iterator>::value_type;
 
-  typedef std::pair<int, std::string> atom_id;
+  typedef std::pair<long, std::string> atom_id;
 
   struct FragInfo {
     std::vector<double> _weights;
     std::vector<atom_id> _mapatom_ids;
     std::vector<atom_id> _mdatom_ids;
-    std::vector<int> _map_local_frame;
+    std::vector<long> _map_local_frame;
   };
 
   struct Seginfo {
-    std::pair<int, int> minmax;
-    std::vector<int> mdatoms;
+    std::pair<long, long> minmax;
+    std::vector<long> mdatoms;
     std::vector<FragInfo> fragments;
     bool map2md;
     std::string segname;
@@ -69,8 +69,8 @@ class SegmentMapper {
   std::map<std::string, std::string> _mapatom_xml;
   std::map<std::string, Seginfo> _segment_info;
 
-  int FindVectorIndexFromAtomId(
-      int atomid, const std::vector<mapAtom*>& fragment_mapatoms) const;
+  long FindVectorIndexFromAtomId(
+      long atomid, const std::vector<mapAtom*>& fragment_mapatoms) const;
 
   void ParseFragment(Seginfo& seginfo, const tools::Property& frag);
 
@@ -86,8 +86,8 @@ class SegmentMapper {
                       const std::vector<const Atom*>& fragment_mdatoms) const;
 
   Logger& _log;
-  std::pair<int, int> CalcAtomIdRange(const Segment& seg) const;
-  std::pair<int, int> CalcAtomIdRange(const std::vector<int>& seg) const;
+  std::pair<long, long> CalcAtomIdRange(const Segment& seg) const;
+  std::pair<long, long> CalcAtomIdRange(const std::vector<long>& seg) const;
 
   atom_id StringToMapIndex(const std::string& map_string) const;
