+++ src/libxtp/calculators/eanalyze.cc
@@ -157,9 +157,9 @@ void EAnalyze::SiteHist(QMStateType state) const {
   double MAX = *std::max_element(Es.begin(), Es.end());
   double MIN = *std::min_element(Es.begin(), Es.end());
   double sum = std::accumulate(Es.begin(), Es.end(), 0.0);
-  double AVG = sum / Es.size();
+  double AVG = sum / double(Es.size());
   double sq_sum = std::inner_product(Es.begin(), Es.end(), Es.begin(), 0.0);
-  double STD = std::sqrt(sq_sum / Es.size() - AVG * AVG);
+  double STD = std::sqrt(sq_sum / double(Es.size()) - AVG * AVG);
 
   // Prepare bins
   int BIN = int((MAX - MIN) / _resolution_sites + 0.5) + 1;
@@ -230,9 +230,9 @@ void EAnalyze::PairHist(const Topology &top, QMStateType state) const {
   double MAX = *std::max_element(dE.begin(), dE.end());
   double MIN = *std::min_element(dE.begin(), dE.end());
   double sum = std::accumulate(dE.begin(), dE.end(), 0.0);
-  double AVG = sum / dE.size();
+  double AVG = sum / double(dE.size());
   double sq_sum = std::inner_product(dE.begin(), dE.end(), dE.begin(), 0.0);
-  double STD = std::sqrt(sq_sum / dE.size() - AVG * AVG);
+  double STD = std::sqrt(sq_sum / double(dE.size()) - AVG * AVG);
   int BIN = int((MAX - MIN) / _resolution_pairs + 0.5) + 1;
 
   std::string filename2 = "eanalyze.pairhist_" + state.ToString() + ".out";
@@ -259,14 +259,15 @@ void EAnalyze::SiteCorr(const Topology &top, QMStateType state) const {
   }
 
   double sum = std::accumulate(Es.begin(), Es.end(), 0.0);
-  double AVG = sum / Es.size();
+  double AVG = sum / double(Es.size());
   double sq_sum = std::inner_product(Es.begin(), Es.end(), Es.begin(), 0.0);
-  double VAR = sq_sum / Es.size() - AVG * AVG;
+  double VAR = sq_sum / double(Es.size()) - AVG * AVG;
   double STD = std::sqrt(VAR);
 
   // Collect inter-site distances, correlation product
   tools::Table tabcorr;
-  int length = _seg_shortlist.size() * (_seg_shortlist.size() - 1) / 2;
+  long unsigned length =
+      _seg_shortlist.size() * (_seg_shortlist.size() - 1) / 2;
   tabcorr.resize(length);
   int index = 0;
   for (unsigned i = 0; i < _seg_shortlist.size(); i++) {
@@ -310,13 +311,14 @@ void EAnalyze::SiteCorr(const Topology &top, QMStateType state) const {
     for (double entry : histCs[bin]) {
       corr += entry / VAR;
     }
-    corr = corr / histCs[bin].size();
+    corr = corr / double(histCs[bin].size());
     for (unsigned i = 0; i < histCs[bin].size(); ++i) {
-      dcorr2 += (histCs[bin][i] / VAR / histCs[bin].size() - corr) *
-                (histCs[bin][i] / VAR / histCs[bin].size() - corr);
+      dcorr2 += (histCs[bin][i] / VAR / double(histCs[bin].size()) - corr) *
+                (histCs[bin][i] / VAR / double(histCs[bin].size()) - corr);
     }
     // error on mean value
-    dcorr2 = dcorr2 / histCs[bin].size() / (histCs[bin].size() - 1);
+    dcorr2 =
+        dcorr2 / double(histCs[bin].size()) / double(histCs[bin].size() - 1);
     double R = MIN + bin * _resolution_space;
     histC.set(bin, R, corr, ' ', std::sqrt(dcorr2));
   }
