+++ src/libxtp/calculators/ianalyze.cc
@@ -135,11 +135,11 @@ void IAnalyze::IHist(Topology &top, QMStateType state) {
   double MAX = *std::max_element(J2s.begin(), J2s.end());
   double MIN = *std::min_element(J2s.begin(), J2s.end());
   double sum = std::accumulate(J2s.begin(), J2s.end(), 0.0);
-  double AVG = sum / J2s.size();
+  double AVG = sum / double(J2s.size());
   double sq_sum = std::inner_product(J2s.begin(), J2s.end(), J2s.begin(), 0.0);
-  double STD = std::sqrt(sq_sum / J2s.size() - AVG * AVG);
+  double STD = std::sqrt(sq_sum / double(J2s.size()) - AVG * AVG);
   // Prepare bins
-  int BIN = ((MAX - MIN) / _resolution_logJ2 + 0.5) + 1;
+  int BIN = int((MAX - MIN) / _resolution_logJ2 + 0.5) + 1;
 
   tools::HistogramNew hist;
   hist.Initialize(MIN, MAX, BIN);
@@ -182,7 +182,7 @@ void IAnalyze::IRdependence(Topology &top, QMStateType state) {
   double MINR = *std::min_element(distances.begin(), distances.end());
 
   // Prepare R bins
-  int pointsR = (MAXR - MINR) / _resolution_space;
+  int pointsR = int((MAXR - MINR) / _resolution_space);
   std::vector<std::vector<double> > rJ2;
   rJ2.resize(pointsR);
 
@@ -206,11 +206,11 @@ void IAnalyze::IRdependence(Topology &top, QMStateType state) {
   for (unsigned i = 0; i < rJ2.size(); i++) {
     const std::vector<double> &vec = rJ2[i];
     double sum = std::accumulate(vec.begin(), vec.end(), 0.0);
-    double AVG = sum / vec.size();
+    double AVG = sum / double(vec.size());
     double thisR = MINR + (i + 0.5) * _resolution_space;
     double sq_sum =
         std::inner_product(vec.begin(), vec.end(), vec.begin(), 0.0);
-    double STD = std::sqrt(sq_sum / vec.size() - AVG * AVG);
+    double STD = std::sqrt(sq_sum / double(vec.size()) - AVG * AVG);
     tab.set(i, thisR, AVG, ' ', STD);
   }
   std::string filename = "ianalyze.ispatial_" + state.ToString() + ".out";
